+++ GPU (CUDA)/src/vector.cu
@@ -1,6 +1,6 @@
-#include <iostream>
-#include <cmath>
 #include "vector.h"
+#include <cmath>
+#include <iostream>
 
 __host__ __device__ Vec::Vec()
 {
@@ -8,7 +8,7 @@ __host__ __device__ Vec::Vec()
     y = 0;
     z = 0;
 }
-__host__ __device__ Vec::Vec(double _x, double _y, double _z)
+__host__ __device__ Vec::Vec(float _x, float _y, float _z)
 {
     x = _x;
     y = _y;
@@ -37,11 +37,11 @@ __host__ __device__ Vec operator*(Vec u, Vec v)
     return Vec(u.x * v.x, u.y * v.y, u.z * v.z);
 }
 
-__host__ __device__ Vec Vec::operator*(double k)
+__host__ __device__ Vec Vec::operator*(float k)
 {
     return Vec(x * k, y * k, z * k);
 }
-__host__ __device__ Vec Vec::operator/(double k)
+__host__ __device__ Vec Vec::operator/(float k)
 {
     return Vec(x / k, y / k, z / k);
 }
@@ -49,20 +49,20 @@ __host__ __device__ Vec Vec::Cross(Vec u, Vec v)
 {
     return Vec(u.y * v.z - u.z * v.y, u.z * v.x - u.x * v.z, u.x * v.y - u.y * v.x);
 }
-__host__ __device__ double Vec::Mod()
+__host__ __device__ float Vec::Mod()
 {
     return sqrt(x * x + y * y + z * z);
 }
-__host__ __device__ double Vec::ModSq()
+__host__ __device__ float Vec::ModSq()
 {
     return x * x + y * y + z * z;
 }
 __host__ __device__ Vec Vec::Norm()
 {
-    double size = sqrt(x * x + y * y + z * z);
+    float size = sqrt(x * x + y * y + z * z);
     return Vec(x / size, y / size, z / size);
 }
-__host__ __device__ double Vec::Dot(Vec u, Vec v)
+__host__ __device__ float Vec::Dot(Vec u, Vec v)
 {
     return u.x * v.x + u.y * v.y + u.z * v.z;
 }
